+++ steamGamesStarter@KopfdesDaemons/files/steamGamesStarter@KopfdesDaemons/helpers/steam.js
@@ -8,6 +8,9 @@ const FILTERED_APP_IDS = [
   "1628350", // Proton Experimental
   "1493710", // Steam Linux Runtime 3.0 (sniper)
   "1391110", // Steam Linux Runtime 2.0 (soldier)
+  "1826330", // Proton EasyAntiCheat Runtime
+  "2348590", // Proton 8.0
+  "2805730", // Proton 9.0
 ];
 
 var SteamHelper = class SteamHelper {
@@ -86,7 +89,7 @@ var SteamHelper = class SteamHelper {
     }
 
     const games = await Promise.all(gamePromises);
-    const filteredGames = games.filter(game => game !== null && game.lastPlayed && !FILTERED_APP_IDS.includes(game.appid));
+    const filteredGames = games.filter(game => game !== null && !FILTERED_APP_IDS.includes(game.appid));
 
     // Filter and sort the games by last played date (newest first)
     const sortedGames = filteredGames.sort((a, b) => parseInt(b.lastPlayed, 10) - parseInt(a.lastPlayed, 10));